Refactor of the Backend (before server code)
