+++ backend/model/CloudFunctions.js
@@ -1,0 +1,9 @@
+// Cloud Functions that will reset User values at the end of the day
+var firebase = require("../firebase/config");
+require("firebase/firestore");
+require("firebase/auth");
+
+// Firestore
+
+// Export the module
+module.exports = {}; // awaiting to be filled
