+++ backend/model/Corkboard.js
@@ -1,0 +1,10 @@
+// Corkboard file for Creating, Reading, Updating, and
+// Deleting corkboard hugs and corkboard Management
+var firebase = require("../firebase/config");
+require("firebase/firestore");
+require("firebase/auth");
+
+// Firestore
+
+// Export the module
+module.exports = {}; // awaiting to be filled
