+++ backend/model/Friends.js
@@ -1,0 +1,10 @@
+// Friends file for Creating, Reading, Updating, and Deleting
+// Friends and Friend Management
+var firebase = require("../firebase/config");
+require("firebase/firestore");
+require("firebase/auth");
+
+// Firestore
+
+// Export the module
+module.exports = {}; // awaiting to be filled
