+++ backend/model/Notifications.js
@@ -1,0 +1,10 @@
+// Notifications file for Creating, Reading, Updating, and Deleting
+// Notifications and Requests
+var firebase = require("../firebase/config");
+require("firebase/firestore");
+require("firebase/auth");
+
+// Firestore
+
+// Export the module
+module.exports = {}; // awaiting to be filled
