--- backend/routes/Corkboard.js
@@ -1,6 +1,0 @@
-// Corkboard file for Creating, Reading, Updating, and
-// Deleting corkboard hugs and corkboard Management
-import Fire from "../firebase/config";
-
-// Firestore
-export const db = Fire.firestore();
