--- backend/routes/Friends.js
@@ -1,6 +1,0 @@
-// Friends file for Creating, Reading, Updating, and Deleting
-// Friends and Friend Management
-import Fire from "../firebase/config";
-
-// Firestore
-export const db = Fire.firestore();
