--- backend/routes/Hugs.js
@@ -1,9 +1,0 @@
-// Hugs file for Creating, Reading, Updating, and Deleting Hugs
-import Fire from "../firebase/config";
-
-// Firestore
-export const db = Fire.firestore();
-
-const HugsAPI = {
-    createHug(friend_id, message, image)
-}
