--- backend/routes/LogIn.js
@@ -1,46 +1,0 @@
-import Fire from "../firebase/config";
-import "firebase/firestore";
-import "firebase/auth";
-
-const LoginAPI = {
-  loginUser: async function(email, password) {
-    let loggedin = false;
-    await Fire.auth()
-    .signInWithEmailAndPassword(email, password)
-    .then(() => {
-      loggedin = true;
-    })
-    .catch(function (error) {
-      loggedin = false;
-      var errorCode = error.code;
-      var errorMessage = error.message;
-
-      switch (error.code) {
-        case "auth/invalid-email":
-          alert("Invalid email.");
-          break;
-
-        case "auth/user-disabled":
-          alert("User disabled.");
-          break;
-
-        case "auth/user-not-found":
-          alert("User not found.");
-          break;
-
-        case "auth/wrong-password":
-          alert("Incorrect password.");
-          break;
-
-        default:
-          alert(errorMessage);
-          break;
-      }
-
-      console.log(error);
-    })
-    return loggedin;
-  }
-}
-
-export default LoginAPI;
