--- backend/routes/Notifications.js
@@ -1,6 +1,0 @@
-// Notifications file for Creating, Reading, Updating, and Deleting
-// Notifications and Requests
-import Fire from "../firebase/config";
-
-// Firestore
-export const db = Fire.firestore();
