+++ backend/test/main.js
@@ -1,0 +1,9 @@
+// Main.js used to run and test functions
+var Users = require("../model/Users");
+
+// Users
+Users.UsersAPI.addUser("username", "first", "last");
+
+// Friends
+
+// Hugs
